+++ src/components/chrome/ChromeFields.js
@@ -113,12 +113,16 @@ export class ChromeFields extends ReactCSS.Component {
 
   handleChange = (data: any) => {
     if (data.hex) {
-      color.isValidHex(data.hex) && this.props.onChange(data.hex)
+      color.isValidHex(data.hex) && this.props.onChange({
+        hex: data.hex,
+        source: 'hex',
+      })
     } else if (data.r || data.g || data.b) {
       this.props.onChange({
         r: data.r || this.props.rgb.r,
         g: data.g || this.props.rgb.g,
         b: data.b || this.props.rgb.b,
+        source: 'rgb',
       })
     } else if (data.a) {
       if (data.a < 0) {
@@ -132,6 +136,7 @@ export class ChromeFields extends ReactCSS.Component {
         s: this.props.hsl.s,
         l: this.props.hsl.l,
         a: Math.round(data.a * 100) / 100,
+        source: 'rgb',
       })
     } else if (data.h || data.s || data.l) {
 
@@ -139,6 +144,7 @@ export class ChromeFields extends ReactCSS.Component {
         h: data.h || this.props.hsl.h,
         s: data.s && (data.s).replace('%', '') || this.props.hsl.s,
         l: data.l && (data.l).replace('%', '') || this.props.hsl.l,
+        source: 'hsl',
       })
     }
   }