+++ src/components/common/Alpha.js
@@ -70,7 +70,13 @@ export class Alpha extends ReactCSS.Component {
     }
 
     if (this.props.a !== a) {
-      this.props.onChange({ h: this.props.hsl.h, s: this.props.hsl.s, l: this.props.hsl.l, a: a })
+      this.props.onChange({
+        h: this.props.hsl.h,
+        s: this.props.hsl.s,
+        l: this.props.hsl.l,
+        a: a,
+        source: 'rgb',
+      })
     }
   }
 
@@ -102,7 +108,9 @@ export class Alpha extends ReactCSS.Component {
           <Checkboard />
         </div>
         <div is="gradient" />
-        <div is="container" ref="container" onMouseDown={ this.handleMouseDown } onTouchMove={ this.handleChange }>
+        <div is="container" ref="container" onMouseDown={ this.handleMouseDown }
+            onTouchMove={ this.handleChange }
+            onTouchStart={ this.handleChange }>
           <div is="pointer" ref="pointer">
             { pointer }
           </div>