+++ src/components/common/Hue.js
@@ -72,7 +72,13 @@ export class Hue extends ReactCSS.Component {
       }
 
       if (this.props.hsl.h !== h) {
-        this.props.onChange({ h: h, s: this.props.hsl.s, l: this.props.hsl.l, a: this.props.hsl.a })
+        this.props.onChange({
+          h: h,
+          s: this.props.hsl.s,
+          l: this.props.hsl.l,
+          a: this.props.hsl.a,
+          source: 'rgb',
+        })
       }
     } else {
       var h
@@ -86,7 +92,13 @@ export class Hue extends ReactCSS.Component {
       }
 
       if (this.props.hsl.h !== h) {
-        this.props.onChange({ h: h, s: this.props.hsl.s, l: this.props.hsl.l, a: this.props.hsl.a })
+        this.props.onChange({
+          h: h,
+          s: this.props.hsl.s,
+          l: this.props.hsl.l,
+          a: this.props.hsl.a,
+          source: 'rgb',
+        })
       }
     }
   }
@@ -115,7 +127,9 @@ export class Hue extends ReactCSS.Component {
 
     return (
       <div is="hue">
-        <div is="container" ref="container" onMouseDown={ this.handleMouseDown } onTouchMove={ this.handleChange }>
+        <div is="container" ref="container" onMouseDown={ this.handleMouseDown }
+            onTouchMove={ this.handleChange }
+            onTouchStart={ this.handleChange }>
           <div is="pointer" ref="pointer">
             { pointer }
           </div>