+++ src/components/common/Saturation.js
@@ -76,7 +76,13 @@ export class Saturation extends ReactCSS.Component {
     var saturation = left * 100 / containerWidth
     var bright = -(top * 100 / containerHeight) + 100
 
-    this.throttle(this.props.onChange, { h: this.props.hsl.h, s: saturation, v: bright, a: this.props.hsl.a })
+    this.throttle(this.props.onChange, {
+      h: this.props.hsl.h,
+      s: saturation,
+      v: bright,
+      a: this.props.hsl.a,
+      source: 'rgb',
+    })
   }
 
   handleMouseDown = (e: any) => {
@@ -102,7 +108,9 @@ export class Saturation extends ReactCSS.Component {
     }
 
     return (
-      <div is="color" ref="container" onMouseDown={ this.handleMouseDown } onTouchMove={ this.handleChange }>
+      <div is="color" ref="container" onMouseDown={ this.handleMouseDown }
+          onTouchMove={ this.handleChange }
+          onTouchStart={ this.handleChange }>
         <div is="white">
           <div is="black" />
           <div is="pointer" ref="pointer">