+++ src/components/photoshop/PhotoshopFields.js
@@ -93,18 +93,23 @@ export class PhotoshopPicker extends ReactCSS.Component {
 
   handleChange = (data: any) => {
     if (data['#']) {
-      color.isValidHex(data['#']) && this.props.onChange(data['#'])
+      color.isValidHex(data['#']) && this.props.onChange({
+        hex: data['#'],
+        source: 'hex',
+      })
     } else if (data.r || data.g || data.b) {
       this.props.onChange({
         r: data.r || this.props.rgb.r,
         g: data.g || this.props.rgb.g,
         b: data.b || this.props.rgb.b,
+        source: 'rgb',
       })
     } else if (data.h || data.s || data.v) {
       this.props.onChange({
         h: data.h || this.props.hsv.h,
         s: data.s || this.props.hsv.s,
         v: data.v || this.props.hsv.v,
+        source: 'hsv',
       })
     }
   }