+++ test/photoshop/PhotoshopFields.test.js
@@ -14,11 +14,10 @@ describe('PhotoshopFields', () => {
 
   it('should pass up hex data onChange', () => {
     props.onChange = chai.spy((data) => {
-      expect(data).to.equal('#333')
+      expect(data.hex).to.equal('#333')
     })
     const PhotoshopFields = TestUtils.renderIntoDocument(<PhotoshopFieldsComponent {...props} />)
     PhotoshopFields.handleChange({ '#': '#333' })
     expect(props.onChange).to.have.been.called
   })
-
 })