+++ test/slider/SliderSwatch.test.js
@@ -18,6 +18,7 @@ describe('SliderSwatch', () => {
         h: 150,
         s: 0.5,
         l: ".50",
+        source: 'hsl',
       }))
     })
     const SliderSwatch = TestUtils.renderIntoDocument(<SliderSwatchComponent {...props} offset=".50" />)
@@ -31,11 +32,11 @@ describe('SliderSwatch', () => {
         h: 150,
         s: 0.5,
         l: ".50",
+        source: 'hsl',
       }))
     })
     const SliderSwatch = TestUtils.renderIntoDocument(<SliderSwatchComponent {...props} offset=".50" />)
     let swatch = SliderSwatch.refs.swatch
     TestUtils.Simulate.click(swatch)
   })
-
 })