+++ test/swatches/Swatches.test.js
@@ -14,14 +14,14 @@ describe('Swatches', () => {
 
   it('should pass up data onChange', () => {
     props.onChange = chai.spy((data) => {
-      expect(data).to.equal('#333')
+      expect(data.hex).to.equal('#333')
     })
     const SwatchesComponent = TestUtils.renderIntoDocument(<Swatches {...props} />)
     SwatchesComponent.handleChange('#333')
     expect(props.onChange).to.have.been.called
   })
 
-  it('should render SwatchesGroup for each top-level array passed to props.colors', () => {
+  it('should render SwatchesGoup for each top-level array passed to props.colors', () => {
     const SwatchesComponent = TestUtils.renderIntoDocument(<Swatches {...props} colors={[['#333'], ['#fff'], ['#aaa'], ['#ddd']]} />)
     let groups = SwatchesComponent.refs.body._reactInternalComponent._renderedChildren
     let groupCount = 0